refactor(sandbox): extract detection constants to sandbox/detection.rs
Move TOOLCHAIN_PASSTHROUGH_CANDIDATES, SENSITIVE_BUT_USEFUL_CANDIDATES,
CLAUDE_AUTO_COPY_CANDIDATE, and the three detect_*() functions out of
dotfiles.rs into a new detection.rs module. The setup helpers in
cli/commands/setup/helpers.rs now import from sandbox::detection.

No functional changes — pure code move with 11 tests relocated.
